restyle add-transaction component using tailwind
