+++ web-app/src/app/pages/cryptos/components/add-transaction/add-transaction.component.html
@@ -1,10 +1,10 @@
 <form class="space-y-4 p-6 bg-white dark:bg-gray-800 rounded-md shadow-md border border-gray-300 dark:border-gray-700"
   [formGroup]="transactionForm" (ngSubmit)="save()">
-  <h1 class="text-2xl font-bold text-gray-900 dark:text-gray-100 mb-4">Add transaction</h1>
+  <h1 class="text-xl font-semibold text-gray-900 dark:text-gray-100">New Order</h1>
 
   <div class="space-y-2">
     <label for="transactionType" class="block text-sm font-medium text-gray-700 dark:text-gray-300">
-      Transaction Type
+      Type
     </label>
     <select id="transactionType" name="transactionType" formControlName="transactionType"
       class="w-full px-4 py-2 text-gray-900 bg-white dark:bg-gray-700 dark:text-gray-100 border border-gray-300 dark:border-gray-600 rounded-lg focus:ring-2 focus:ring-blue-500 focus:border-blue-500 transition-colors duration-200">
@@ -39,7 +39,7 @@
 
   <div class="space-y-2">
     <label for="purchaseDate" class="block text-sm font-medium text-gray-700 dark:text-gray-300">
-      Date of Purchase
+      Order Date
     </label>
     <input type="date" id="purchaseDate" name="purchaseDate" formControlName="purchaseDate"
       class="w-full px-4 py-2 text-gray-900 bg-white dark:bg-gray-700 dark:text-gray-100 border border-gray-300 dark:border-gray-600 rounded-lg focus:ring-2 focus:ring-blue-500 focus:border-blue-500 transition-colors duration-200">
