+++ web-app/src/app/pages/cryptos/components/stats-card/stats-card.component.html
@@ -8,7 +8,7 @@
       <span class="text-base font-normal text-gray-900 dark:text-gray-100">{{ value() }}</span>
       }
       @if (percentDifference()) {
-      <app-percent-difference [percentDifference]="percentDifference()">
+      <app-percent-difference class="ml-1" [percentDifference]="percentDifference()">
       </app-percent-difference>
       }
     </div>
