+++ web-app/src/app/pages/cryptos/components/transaction-table/transaction-table.component.html
@@ -1,0 +1,67 @@
+<div class="p-6 mb-4 bg-white dark:bg-gray-800 rounded-md shadow-md border border-gray-300 dark:border-gray-700">
+  <header class="flex justify-between items-center mb-6 sm:flex-col sm:items-start">
+    <h1 class="text-xl font-semibold text-gray-900 dark:text-gray-100">Orders</h1>
+    <div class="flex items-center space-x-2">
+      <label for="filterByDate" class="text-sm font-medium text-gray-700 dark:text-gray-300">
+        Order by
+      </label>
+      <select id="filterByDate"
+        class="px-4 py-2 text-sm text-gray-900 bg-white dark:bg-gray-700 dark:text-gray-100 border border-gray-300 dark:border-gray-600 rounded-lg focus:ring-2 focus:ring-blue-500 focus:border-blue-500 transition-colors duration-200">
+        <option *ngFor="let option of orderOptions" [value]="option">{{ option }}</option>
+      </select>
+    </div>
+  </header>
+
+  <div class="overflow-x-auto">
+    <table class="min-w-full divide-y divide-gray-200 dark:divide-gray-700">
+      <thead class="bg-gray-50 dark:bg-gray-700">
+        <tr>
+          <th class="px-6 py-3 text-left text-xs font-medium text-gray-500 dark:text-gray-300 uppercase tracking-wider">
+            Amount</th>
+          <th class="px-6 py-3 text-left text-xs font-medium text-gray-500 dark:text-gray-300 uppercase tracking-wider">
+            Price Unit</th>
+          <th class="px-6 py-3 text-left text-xs font-medium text-gray-500 dark:text-gray-300 uppercase tracking-wider">
+            %</th>
+          <th class="px-6 py-3 text-left text-xs font-medium text-gray-500 dark:text-gray-300 uppercase tracking-wider">
+            Type</th>
+          <th class="px-6 py-3 text-left text-xs font-medium text-gray-500 dark:text-gray-300 uppercase tracking-wider">
+            Date</th>
+          <th class="px-6 py-3 text-left text-xs font-medium text-gray-500 dark:text-gray-300 uppercase tracking-wider">
+            Exchange</th>
+        </tr>
+      </thead>
+      <tbody class="bg-white dark:bg-gray-800 divide-y divide-gray-200 dark:divide-gray-700">
+        <tr *ngFor="let element of transactions | async trackBy: transactionById"
+          class="hover:bg-gray-50 dark:hover:bg-gray-700 transition-colors duration-150">
+          <td class="px-6 py-4 whitespace-nowrap text-sm text-gray-900 dark:text-gray-100">
+            {{ element.amount }}
+          </td>
+          <td class="px-6 py-4 whitespace-nowrap text-sm text-gray-900 dark:text-gray-100">
+            {{ element.price | formatCurrency }}
+          </td>
+          <td class="px-6 py-4 whitespace-nowrap text-sm">
+            <app-percent-difference *ngIf="element.transactionType === 1"
+              [percentDifference]="element.percentDifference">
+            </app-percent-difference>
+          </td>
+          <td class="px-6 py-4 whitespace-nowrap text-sm">
+            <span *ngIf="element.transactionType === 1"
+              class="inline-flex items-center px-2.5 py-0.5 rounded-full text-xs font-medium bg-green-100 text-green-800 dark:bg-green-700 dark:text-green-100">
+              Buy
+            </span>
+            <span *ngIf="element.transactionType === 2"
+              class="inline-flex items-center px-2.5 py-0.5 rounded-full text-xs font-medium bg-blue-100 text-blue-800 dark:bg-blue-700 dark:text-blue-100">
+              Sell
+            </span>
+          </td>
+          <td class="px-6 py-4 whitespace-nowrap text-sm text-gray-900 dark:text-gray-100">
+            {{ element.purchaseDate | date: 'dd/MM/y' }}
+          </td>
+          <td class="px-6 py-4 whitespace-nowrap text-sm text-gray-900 dark:text-gray-100">
+            {{ element.exchangeName }}
+          </td>
+        </tr>
+      </tbody>
+    </table>
+  </div>
+</div>
