+++ web-app/src/app/pages/cryptos/components/transaction-table/transaction-table.component.ts
@@ -1,0 +1,34 @@
+import { Component, Input } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { FormsModule } from '@angular/forms';
+import { CryptoTransactionHistory } from 'src/app/core/models/crypto-transaction-history';
+import { BehaviorSubject } from 'rxjs';
+import { PercentDifferenceComponent } from '../percent-difference.component';
+import { FormatCurrencyPipe } from 'src/app/core/pipes/format-currency.pipe';
+
+@Component({
+  selector: 'app-transaction-table',
+  standalone: true,
+  imports: [
+    CommonModule,
+    FormsModule,
+    PercentDifferenceComponent,
+    FormatCurrencyPipe],
+  templateUrl: './transaction-table.component.html',
+})
+export class TransactionTableComponent {
+  orderOptions: any[] = [
+    'Amount (Low to High)',
+    'Amount (High to Low)',
+    'Price Unit (Low to High)',
+    'Price Unit (High to Low)',
+    'Order Type (Buy First)',
+    'Order Type (Sell First)',
+    'Exchange Name (A-Z)',
+    'Exchange Name (Z-A)'
+  ]
+  @Input() transactions: BehaviorSubject<CryptoTransactionHistory[]> = new BehaviorSubject<CryptoTransactionHistory[]>([]);
+  transactionById(index: number, transaction: CryptoTransactionHistory) {
+    return transaction.id;
+  }
+}
