+++ web-app/src/app/pages/cryptos/containers/crypto-details/crypto-details.component.html
@@ -1,4 +1,4 @@
-<div class="max-w-6xl mx-auto px-4 sm:px-4 lg:px-8 py-4">
+<div class="mx-auto p-6">
   <h1 class="text-2xl font-bold text-gray-900 dark:text-gray-100 mb-4">
     {{ cryptoInfo.symbol | uppercase }}
   </h1>
