+++ web-app/src/app/pages/cryptos/containers/view-cryptos/view-cryptos.component.html
@@ -44,7 +44,7 @@
                 [setBalanceStatus]="setBalanceStatus">
               </app-crypto-filter>
             </div>
-            <div class="mr-0 inline-block">
+            <div class="inline-block ml-1">
               <app-add-crypto (cryptoCreated)="loadCryptoAssets()"></app-add-crypto>
             </div>
           </div>
